services/resume-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
spacy==3.7.2
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0
boto3==1.34.0
python-multipart==0.0.6
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.25.2
opencv-python==4.8.1.78
Pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
hypothesis==6.88.1